Adding StopSearchTimeSeries
