+++ traffic_stops/static/js/app/states/nc/StopSearch.js
@@ -1,16 +1,5 @@
-import AggregateDataHandlerBase from '../../base/AggregateDataHandlerBase.js';
-import TableBase from '../../base/TableBase.js';
-import { StopRatioTimeSeries } from './Stops.js';
 import Stops from './defaults.js';
-
 import * as C from '../../common/StopSearch.js';
-
-import _ from 'underscore';
-import d3 from 'd3';
-import Backbone from 'backbone';
-import $ from 'jquery';
-
-Backbone.$ = $;
 
 const StopSearchHandler = C.StopSearchHandlerBase.extend({
   Stops: Stops,
@@ -22,61 +11,32 @@ const StopSearchHandler = C.StopSearchHandlerBase.extend({
   _pprint: (d) => Stops.pprint.get(d)
 });
 
-var StopSearchTimeSeries = StopRatioTimeSeries.extend({
-  setDefaultChart: function(){
-    this.chart = nv.models.lineChart()
-                  .useInteractiveGuideline (true)
-                  .transitionDuration(350)
-                  .showLegend(true)
-                  .showYAxis(true)
-                  .showXAxis(true)
-                  .width(this.get("width"))
-                  .height(this.get("height"));
-
-    this.chart.xAxis
-        .axisLabel('Year');
-
-    this.chart.yAxis
-        .tickFormat(d3.format('%'));
+const StopSearchTimeSeries = C.StopSearchTimeSeriesBase.extend({
+  defaults: {
+    showEthnicity: false,
+    width: 750,
+    height: 375
   },
-  _formatData: function(){
 
-    var data = StopSearchTimeSeries.__super__._formatData.call(this),
-        total = this.data.line.get('total'),
-        defaultEnabled = ["Total", "White", "Black", "Hispanic", "Non-hispanic"];
+  Stops: Stops,
 
-    // add total-line
-    data.push({
-      key: "Total",
-      values: total,
-      color: Stops.baseline_color,
-      disabled: false
-    });
+  defaultEnabled: ["Total", "White", "Black", "Hispanic", "Non-hispanic"],
 
-    // predefine enabled/disabled lines on chart
-    _.each(data, function(d){
-      d.disabled = defaultEnabled.indexOf(d.key) === -1;
-    });
-
-    return data;
+  _items: function () {
+    return (this.get('showEthnicity')) ? Stops.ethnicities : Stops.races;
   },
-});
 
-var StopSearchTable = TableBase.extend({
-  get_tabular_data: function(){
-    var header, row, rows = [];
+  _pprint: function (type) {
+    return Stops.pprint.get(type);
+  },
 
-    // create header
-    rows.push(this.data.table_headers);
-
-    // create data rows
-    _.each(this.data.table, function(k, v){
-      rows.push(k)
-    });
-
-    return rows;
+  triggerRaceToggle: function(e, v){
+    this.set('showEthnicity', v);
+    this.drawChart();
   }
 });
+
+var StopSearchTable = C.StopSearchTableBase.extend({});
 
 export default {
   StopSearchHandler,
